feature/5_AngularMaterial

+++ DynamicFilter.Domain.Core/Models/AssistantResultReportModel.cs
@@ -1,6 +1,6 @@
 ﻿namespace DynamicFilter.Domain.Core.Models {
     public class AssistantResultReportModel {
         public Item Item { get; set; }
-        public float RateInPercent { get; set; }
+        public decimal RateInPercent { get; set; }
     }
 }
